+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## v1.1
 
-- refactored `userChrome.css`
+- changed `userChrome.css` to use a hidden attribute in a tab's favicon
 
 ## v1.0